[all] Switch to pedantic analysis package (#1497)
Switches our out of date fork of Flutter's analysis rules to the
pedantic package. In order to make this an easy change to initially land
and then incrementally enable, any existing failures have just been
ignored within their respective packages instead of being fixed.

Also adds a global lint for missing public DartDocs. Like the pedantic
lints, this is ignored in packages that are already failing it.

This is a mirror of a similar change to flutter/plugins, #2300.